add router and info component
