+++ client/src/Pages/Info/Info.js
@@ -1,0 +1,13 @@
+import React, {useState, useEffect} from 'react';
+
+const Info = () => {
+    return(
+        <div>
+            <h1>Info</h1>
+        </div>
+    );
+}
+
+
+
+export default Info;
